+++ src/components/GoodIcon.tsx
